% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{report_app}
\alias{report_app}
\title{Alias for [run_app()]}
\usage{
report_app(res = NULL, launch.browser = TRUE, ...)
}
\arguments{
\item{res}{A PM_result object.}

\item{launch.browser}{Passed to [shiny::runApp()].}

\item{...}{Additional options passed into golem options.}
}
\description{
Alias for [run_app()]
}
