% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Pmetrics Reports app}
\usage{
run_app(res = NULL, launch.browser = TRUE, ...)
}
\arguments{
\item{res}{A PM_result object.}

\item{launch.browser}{Passed to [shiny::runApp()].}

\item{...}{Additional options passed into golem options.}
}
\description{
Launch the Pmetrics Reports app
}
